@@ -7,17 +7,15 @@
     <packaging>pom</packaging>
     <name>${project.artifactId}</name>
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <spring.boot.version>2.3.1.RELEASE</spring.boot.version>
     </properties>
     <dependencyManagement>
         <dependencies>
             <dependency>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-starter-web</artifactId>
-                <version>${spring.boot.version}</version>
             </dependency>
 
             <dependency>
@@ -32,15 +30,10 @@
 
     <build>
         <plugins>
-            <!-- TODO check what to do with this plugin -->
-            <!-- <plugin>-->
-            <!--   <groupId>org.springframework.boot</groupId>-->
-            <!--   <artifactId>spring-boot-maven-plugin</artifactId>-->
-            <!-- </plugin>-->
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>2.0.2</version>
+                <version>3.11.0</version>
                 <configuration>
                     <source>${java.version}</source>
                     <target>${java.version}</target>
