@@ -1,10 +1,14 @@
 package be.butskri.test.backwardscompatibility.annotated;
 
-import javax.persistence.Entity;
+import jakarta.persistence.Entity;
 
 @Entity
 public class SomeEntity {
     private SomeEntityField someEntityField;
+
+    protected SomeEntity() {
+        // necessary for JPA
+    }
 
     public SomeEntity(SomeEntityField someEntityField) {
         this.someEntityField = someEntityField;
