@@ -1,7 +1,8 @@
 package be.butskri.test.backwardscompatibility.annotated;
 
-import javax.persistence.Entity;
+import jakarta.persistence.Entity;
 
 @Entity
 public class SomeOtherEntity {
+
 }
